% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetlist.R
\name{datasetlist}
\alias{datasetlist}
\title{Get the list of current data sets and their resources}
\usage{
datasetlist()

datasetlist(topic = c("Administration", "City and Services",
                      "Economy and Business", "Population",
                      "Territory"))

datasetlist(subtopic = c("Culture and Leisure", "Demography",
                         "Education", "Employment", "Environment",
                         "Housing", "Human resources",
                         "Legislation and justice", "Participation",
                         "Procurement", "Public opinion",
                         "Public sector", "Science and technology",
                         "Security", "Society and Welfare", "Sport",
                         "Tourism", "Town planning and Infrastructures",
                         "Trade", "Transport"))
}
\arguments{
\item{topic}{character}

\item{subtopic}{character}
}
\description{
Use \code{datasetlist()} to manage the catalogue of data sets
on the \href{https://opendata-ajuntament.barcelona.cat/en/}{Open Data BCN} website.
This function may have an interest for citizens to know what information is available on the
portal and the data sets that are being added.
}
\details{
\itemize{
 \item{If you want to retrieve all the current data sets
 and their resources, execute the function without arguments}
 \item{You can filter by topic and obtain a subset.}
 \item{If you are having difficulties finding a certain data set,
you can also filter by more specific topics}
}
}
\examples{
# Retrieves all the current data sets
datasetlist()

# Only shows the data sets about population
datasetlist(topic = "Population")

# Only shows the data sets about sports
datasetlist(subtopic = "Sport")
}
\author{
Xavier Vivancos
}
