% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.csv.R
\name{get.csv}
\alias{get.csv}
\title{Retrieve CSV files from the Open Data BCN portal}
\usage{
get.csv(id)
}
\arguments{
\item{id}{character}
}
\description{
Use \code{get.csv()} to extract any CSV data set available on the
\href{https://opendata-ajuntament.barcelona.cat/en/}{Open Data BCN website} and
load it directly in your RStudio environment.

Previously, use \code{\link[baRcelona]{datasetlist}} to get the resource ID
of the CSV you want to consult, and then introduce it as an input argument
using the \code{get.csv()} function.
}
\examples{
# Get the resource ID of the CSV you want to consult
datasetlist()

# Retrieve the CSV
get.csv(id = "e9d8d056-f1c9-4d38-b7ae-aedc9281417e")
}
\author{
Xavier Vivancos
}
