% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCSV.R
\name{getCSV}
\alias{getCSV}
\title{Retrieve CSV files from the Open Data BCN portal}
\usage{
getCSV(id)
}
\arguments{
\item{topic}{character}

\item{subtopic}{character}
}
\description{
Use \code{getCSV()} to read any CSV file available on the
\href{https://opendata-ajuntament.barcelona.cat/en/}{Open Data BCN} website and
load it directly in your RStudio environment. Use \link[baRcelona]{datasetlist}
}
\details{
\itemize{
 \item{If you want to retrieve all the current data sets
 and their resources, execute the function without arguments}
 \item{You can filter by topic and obtain a subset.}
 \item{If you are having difficulties finding a certain data set,
you can also filter by more specific topics}
}
}
\examples{
# Retrieves all the current data sets
datasetlist()

# Only shows the data sets about population
datasetlist(topic = "Population")

# Only shows the data sets about sports
datasetlist(subtopic = "Sport")
}
\author{
Xavier Vivancos
}
