% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getids.R
\name{datasetlist}
\alias{datasetlist}
\title{Get the list of current data sets and their resources}
\usage{
datasetlist()

datasetlist(topic = c("Administration", "City and Services",
                      "Economy and Business", "Population",
                      "Territory"))

datasetlist(subtopic = c("Culture and Leisure", "Demography",
                         "Education", "Employment", "Environment",
                         "Housing", "Human resources", "Sport",
                         "Legislation and justice", "Transport",
                         "Security" , "Public sector", ...))
}
\arguments{
\item{topic}{character}

\item{subtopic}{character}
}
\description{
Use \code{datasetlist()} to manage the catalogue of data sets
on the \href{https://opendata-ajuntament.barcelona.cat/en/}{Open Data BCN} website.
This function may have an interest for citizens to know what information is available on the
portal and the data sets that are being added.
}
\details{
\itemize{
 \item{If you want to retrieve all the current data sets
 and their resources, execute the function without arguments}
 \item{You can filter by topic and obtain a subset.}
 \item{If you are having difficulties finding a certain data set,
you can also filter by more specific topics}
}
}
\examples{
# Retrieve all the current data sets
datasetlist()

# Show only the data set list about population
datasetlist(topic = "Population")
}
\author{
Xavier Vivancos
}
